+++ src/build-tests/index.ts
@@ -2,4 +2,4 @@ export * from './addTarball';
 export * from './cleanup';
 export { THIS1 } from './constants';
 export * from './import';
-export * from './types';
+export * from './imports';